back-end/requirements.txt
alembic==1.7.5
anyio==3.5.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.3
fastapi==0.73.0
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
psycopg2==2.9.3
pydantic==1.9.0
python-dotenv==0.19.2
SQLAlchemy==1.4.31
sqlparse==0.4.1
starlette==0.17.1
itsdangerous==2.0.1
uvicorn
